+++ packages/dropdown-menu/src/util/getDisplayName.js
@@ -1,0 +1,9 @@
+// @flow
+// from https://facebook.github.io/react/docs/higher-order-components.html
+
+// $FlowFixMe
+const getDisplayName = WrappedComponent => (
+  WrappedComponent.displayName || WrappedComponent.name || 'Component'
+);
+
+export default getDisplayName;