+++ packages/editor-core/src/plugins/media/media-links.ts
@@ -1,3 +1,4 @@
+import { Context } from '@atlaskit/media-core';
 import {
   atTheEndOfDoc,
   endPositionOfParent
@@ -8,10 +9,10 @@ import {
   ReplaceStep,
   Transaction,
   Fragment,
-  Mark,
   MarkType,
 } from '../../prosemirror';
 import { posOfMediaGroupBelow, posOfParentMediaGroup } from './utils';
+import { normalizeUrl } from '../hyperlink/utils';
 
 export interface RangeWithUrls {
   start: number;
@@ -19,33 +20,70 @@ export interface RangeWithUrls {
   urls: string[];
 }
 
-export const insertLinks = (view: EditorView, linkRanges: RangeWithUrls[], collection?: string): void => {
-  if (linkRanges.length <= 0 || !collection) {
-    return;
-  }
+export const insertLinks = async (view: EditorView, linkRanges: RangeWithUrls[], linkCreateContext: Context, collection?: string): Promise<Array<string> | undefined> => {
+  return new Promise<Array<string> | undefined>((resolve, reject) => {
+    if (linkRanges.length <= 0 || !collection) {
+      return resolve();
+    }
 
-  const { state, dispatch } = view;
-  const { tr } = state;
+    const { state } = view;
+    const trQueue: Array<Transaction> = [];
+    const links = mapLinksInfo(state, linkRanges);
 
-  const linksInfo = reduceLinksInfo(state, linkRanges);
+    let queueSize = 0;
+    let linkIds: Array<string> = [];
+    const resolveWhenFinished = () => --queueSize === 0 && resolve(linkIds);
 
-  const linkNodes = linksInfo.urls.map((url) => {
-    return state.schema.nodes.media.create({ id: url, type: 'link', collection: collection });
+    links.forEach(({ urls, pos }) => {
+      urls.map(url => {
+        // TODO: Remove the following line once we've solved https://product-fabric.atlassian.net/browse/ED-2321
+        url = normalizeUrl(url);
+
+        queueSize++;
+
+        try {
+          const observable = linkCreateContext.getUrlPreviewProvider(url).observable();
+          observable.subscribe(
+            metadata => {
+              linkCreateContext.addLinkItem(url, collection, metadata).then(id => {
+                const { state, dispatch } = view;
+                const { tr } = state;
+                const node = state.schema.nodes.media.create({ id, type: 'link', collection });
+
+                // If there's multiple replace steps, make sure subsequent transactions are mapped onto new positions
+                trQueue.forEach(tr => {
+                  pos = tr.mapping.map(pos);
+                });
+
+                const $latestPos = tr.doc.resolve(pos);
+                const insertPos = posOfMediaGroupBelow(state, $latestPos, false)
+                  || posOfParentMediaGroup(state, $latestPos, false)
+                  || endPositionOfParent($latestPos);
+
+                // insert an empty paragraph in case we've reached the end of the document
+                if (atTheEndOfDoc(state)) {
+                  tr.insert(insertPos, state.schema.nodes.paragraph.create());
+                }
+
+                tr.replaceWith(insertPos, insertPos, node);
+                trQueue.push(tr);
+                dispatch(tr);
+                linkIds.push(id);
+                resolveWhenFinished();
+              }).catch(() => {
+                resolveWhenFinished();
+              });
+            },
+            e => {
+              resolveWhenFinished();
+            }
+          );
+        } catch(e) {
+          resolveWhenFinished();
+        }
+      });
+    });
   });
-
-  const $latestPos = tr.doc.resolve(linksInfo.latestPos);
-
-  const insertPos = posOfMediaGroupBelow(state, $latestPos, false)
-    || posOfParentMediaGroup(state, $latestPos, false)
-    || endPositionOfParent($latestPos);
-
-  // insert a paragraph after if reach the end of doc
-  if (atTheEndOfDoc(state)) {
-    tr.insert(insertPos, state.schema.nodes.paragraph.create());
-  }
-
-  tr.replaceWith(insertPos, insertPos, linkNodes);
-  dispatch(tr);
 };
 
 export const detectLinkRangesInSteps = (tr: Transaction, link: MarkType): RangeWithUrls[] => {
@@ -54,9 +92,9 @@ export const detectLinkRangesInSteps = (tr: Transaction, link: MarkType): RangeW
   tr.steps.forEach((step) => {
     let rangeWithUrls;
     if (step instanceof AddMarkStep) {
-      rangeWithUrls = findRangesWithUrlsInAddMarkStep(step as AddMarkStep, link);
+      rangeWithUrls = findRangesWithUrlsInAddMarkStep(step, link);
     } else if (step instanceof ReplaceStep) {
-      rangeWithUrls = findRangesWithUrlsInReplaceStep(step as ReplaceStep, link);
+      rangeWithUrls = findRangesWithUrlsInReplaceStep(step, link);
     }
 
     if (rangeWithUrls) {
@@ -67,54 +105,42 @@ export const detectLinkRangesInSteps = (tr: Transaction, link: MarkType): RangeW
   return linkRanges;
 };
 
-const reduceLinksInfo = (state, linkRanges: RangeWithUrls[]): { latestPos: number, urls: string[] } => {
+const mapLinksInfo = (state, linkRanges: RangeWithUrls[]): Array<{ pos: number, urls: string[] }> => {
   const posAtTheEndOfDoc = state.doc.nodeSize - 4;
 
-  const linksInfo = linkRanges.reduce((linksInfo, rangeWithUrl) => {
-    linksInfo.urls = linksInfo.urls.concat(rangeWithUrl.urls);
-    if (rangeWithUrl.end > linksInfo.latestPos) {
-      linksInfo.latestPos = rangeWithUrl.end;
-    }
-
-    if (posAtTheEndOfDoc < linksInfo.latestPos) {
-      linksInfo.latestPos = posAtTheEndOfDoc;
-    }
-    return linksInfo;
-  }, { latestPos: 0, urls: [] as string[] });
-
-  return linksInfo;
+  return linkRanges.map(rangeWithUrls => ({
+    urls: rangeWithUrls.urls,
+    pos: posAtTheEndOfDoc < rangeWithUrls.end ? posAtTheEndOfDoc : rangeWithUrls.end
+  }));
 };
 
 const findRangesWithUrlsInAddMarkStep = (step: AddMarkStep, link: MarkType): RangeWithUrls | undefined => {
   const { mark } = step;
-  if (isValidLinkMark(mark, link)) {
+
+  if (link.isInSet([ mark ]) && mark.attrs.href) {
     return { start: step.from, end: step.to, urls: [mark.attrs.href] };
   }
-};
-
-const isValidLinkMark = (mark: Mark, link: MarkType): boolean => {
-  return mark.type === link && mark.attrs.href.length;
 };
 
 const findRangesWithUrlsInReplaceStep = (step: ReplaceStep, link: MarkType): RangeWithUrls | undefined => {
   const { slice } = step;
   const urls: string[] = findLinksInNodeContent([], slice.content, link);
+
   if (urls.length > 0) {
-    // The end position is step.from + slice.size || step.to is because
-    // it can be replaced by a smaller size content
-    // if simply use step.to, it might be out of range later.
-    return { start: step.from, end: Math.min(step.from + slice.size, step.to), urls: urls };
+    return {
+      start: step.from,
+      end:   step.from + slice.size,
+      urls
+    };
   }
 };
 
 const findLinksInNodeContent = (urls: string[], content: Fragment, link: MarkType) => {
   content.forEach((child) => {
-    const linkMarks = child.marks.filter((mark) => {
-      return isValidLinkMark(mark, link);
-    });
+    const linkMark = link.isInSet(child.marks);
 
-    if (linkMarks.length > 0) {
-      urls.push(linkMarks[0].attrs.href);
+    if (linkMark && linkMark.attrs.href) {
+      urls.push(linkMark.attrs.href);
     }
 
     findLinksInNodeContent(urls, child.content, link);